Updates per Rob's review (yet to do tests)
