+++ opus/import/obs_profile_pds4.py
@@ -52,8 +52,8 @@ class ObsProfilePDS4(ObsProfile, ObsBasePDS4):
     def field_obs_profile_temporal_sampling(self):
         return None
 
-    # Assume BeckerJarmak optical depth parameter can be interpreted the same as the Uranus Occs opacity parameter.
-    # Can fix later.
+    # TODO: Investigate further and fix if necessary. For the moment, assume BeckerJarmak optical depth
+    # parameter can be interpreted the same as the Uranus Occs opacity parameter.
     def field_obs_profile_optical_depth1(self):
         ret = self._index_col('rings:lowest_detectable_opacity') # Uranus Occs
         if ret is None:
