+++ opus/import/obs_volume_coiss_12xxx.py
@@ -9,7 +9,7 @@
 import opus_support
 
 from obs_cassini_common_pds3 import (ObsCassiniCommonPDS3,
-                                       COISS_TARGET_DESC_MAPPING)
+                                     COISS_TARGET_DESC_MAPPING)
 
 
 # Wavelength information for combinations of filters
@@ -354,9 +354,9 @@ class ObsVolumeCOISS12xxx(ObsCassiniCommonPDS3):
         return self._create_mult('NONE')
 
 
-    ############################################
+    ##########################################
     ### OVERRIDE FROM ObsCassiniCommonPDS3 ###
-    ############################################
+    ##########################################
 
     def field_obs_mission_cassini_spacecraft_clock_count1(self):
         partition = self._index_col('SPACECRAFT_CLOCK_CNT_PARTITION')
